+++ CHANGELOG.md
@@ -4,9 +4,13 @@
 
 ---
 
+#### 29.06.2017
+
+* [C] Clock (Exercism)
+* [C] Phone Number (Exercism)
+
 #### 28.06.2017
 
-* [C] Phone Number (Exercism)
 * [C] Atbash Cipher (Exercism)
 * [C] Allergies (Exercism)